+++ app/Http/Controllers/ItemController.php
@@ -137,6 +137,7 @@ class ItemController extends Controller
                 'description_localisation' => 'route-descriptions.item_GET_index',
                 'parameters_config' => 'api.item.parameters.collection',
                 'conditionals' => $this->get_parameters,
+                'pagination' => true,
                 'authenticated' => false
             ],
             [