+++ app/Http/Controllers/ItemSubCategoryController.php
@@ -157,6 +157,7 @@ class ItemSubCategoryController extends Controller
                 'description_localisation' => 'route-descriptions.item_sub_category_GET_index',
                 'parameters_config' => 'api.item-subcategory.parameters.collection',
                 'conditionals' => [],
+                'pagination' => true,
                 'authenticated' => false
             ],
             [