+++ routes/api/public-summary-routes.php
@@ -112,5 +112,15 @@ Route::group(
             'resource_types/{resource_type_id}/resources/{resource_id}/summary/years/{year}/months/{month}',
             'SummaryController@optionsMonth'
         );
+
+        Route::get(
+            'summary/request/access-log/monthly',
+            'SummaryRequestController@monthlyAccessLog'
+        );
+
+        Route::options(
+            'summary/request/access-log/monthly',
+            'SummaryRequestController@optionsMonthlyAccessLog'
+        );
     }
 );